+++ app/src/main/java/com/example/wataritabi/ui/view/destination/Destination.kt
@@ -1,4 +1,4 @@
-package com.example.wataritabi.view.destination
+package com.example.wataritabi.ui.view.destination
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
@@ -26,7 +26,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.example.wataritabi.navigation.BottomBarScreen
-import com.example.wataritabi.viewmodel.MapViewModel
+import com.example.wataritabi.ui.viewmodel.MapViewModel
 import com.google.android.libraries.places.api.Places
 
 @Composable
