+++ app/src/main/java/com/example/wataritabi/ui/view/map/Map.kt
@@ -1,4 +1,4 @@
-package com.example.wataritabi.view.map
+package com.example.wataritabi.ui.view.map
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -17,7 +17,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import com.example.wataritabi.viewmodel.MapViewModel
+import com.example.wataritabi.ui.viewmodel.MapViewModel
 import com.google.android.gms.maps.model.CameraPosition
 import com.google.android.gms.maps.model.LatLng
 import com.google.maps.android.compose.GoogleMap
